+++ airflow/providers/amazon/aws/operators/sagemaker_endpoint_config.py
@@ -16,37 +16,14 @@
 # specific language governing permissions and limitations
 # under the License.
 
-from airflow.exceptions import AirflowException
-from airflow.providers.amazon.aws.operators.sagemaker_base import SageMakerBaseOperator
+"""This module is deprecated. Please use :mod:`airflow.providers.amazon.aws.operators.sagemaker`."""
 
+import warnings
 
-class SageMakerEndpointConfigOperator(SageMakerBaseOperator):
-    """
-    Create a SageMaker endpoint config.
+from airflow.providers.amazon.aws.operators.sagemaker import SageMakerEndpointConfigOperator  # noqa
 
-    This operator returns The ARN of the endpoint config created in Amazon SageMaker
-
-    :param config: The configuration necessary to create an endpoint config.
-
-        For details of the configuration parameter see :py:meth:`SageMaker.Client.create_endpoint_config`
-    :type config: dict
-    :param aws_conn_id: The AWS connection ID to use.
-    :type aws_conn_id: str
-    """
-
-    integer_fields = [['ProductionVariants', 'InitialInstanceCount']]
-
-    def __init__(self, *, config: dict, **kwargs):
-        super().__init__(config=config, **kwargs)
-
-        self.config = config
-
-    def execute(self, context) -> dict:
-        self.preprocess_config()
-
-        self.log.info('Creating SageMaker Endpoint Config %s.', self.config['EndpointConfigName'])
-        response = self.hook.create_endpoint_config(self.config)
-        if response['ResponseMetadata']['HTTPStatusCode'] != 200:
-            raise AirflowException(f'Sagemaker endpoint config creation failed: {response}')
-        else:
-            return {'EndpointConfig': self.hook.describe_endpoint_config(self.config['EndpointConfigName'])}
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.operators.sagemaker`.",
+    DeprecationWarning,
+    stacklevel=2,
+)
