+++ airflow/providers/amazon/aws/operators/sagemaker_model.py
@@ -16,41 +16,14 @@
 # specific language governing permissions and limitations
 # under the License.
 
-from airflow.exceptions import AirflowException
-from airflow.providers.amazon.aws.hooks.base_aws import AwsBaseHook
-from airflow.providers.amazon.aws.operators.sagemaker_base import SageMakerBaseOperator
+"""This module is deprecated. Please use :mod:`airflow.providers.amazon.aws.operators.sagemaker`."""
 
+import warnings
 
-class SageMakerModelOperator(SageMakerBaseOperator):
-    """
-    Create a SageMaker model.
+from airflow.providers.amazon.aws.operators.sagemaker import SageMakerModelOperator  # noqa
 
-    This operator returns The ARN of the model created in Amazon SageMaker
-
-    :param config: The configuration necessary to create a model.
-
-        For details of the configuration parameter see :py:meth:`SageMaker.Client.create_model`
-    :type config: dict
-    :param aws_conn_id: The AWS connection ID to use.
-    :type aws_conn_id: str
-    """
-
-    def __init__(self, *, config, **kwargs):
-        super().__init__(config=config, **kwargs)
-
-        self.config = config
-
-    def expand_role(self) -> None:
-        if 'ExecutionRoleArn' in self.config:
-            hook = AwsBaseHook(self.aws_conn_id, client_type='iam')
-            self.config['ExecutionRoleArn'] = hook.expand_role(self.config['ExecutionRoleArn'])
-
-    def execute(self, context) -> dict:
-        self.preprocess_config()
-
-        self.log.info('Creating SageMaker Model %s.', self.config['ModelName'])
-        response = self.hook.create_model(self.config)
-        if response['ResponseMetadata']['HTTPStatusCode'] != 200:
-            raise AirflowException(f'Sagemaker model creation failed: {response}')
-        else:
-            return {'Model': self.hook.describe_model(self.config['ModelName'])}
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.operators.sagemaker`.",
+    DeprecationWarning,
+    stacklevel=2,
+)
