+++ airflow/providers/amazon/aws/operators/sagemaker_transform.py
@@ -15,108 +15,15 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-from typing import List, Optional
 
-from airflow.exceptions import AirflowException
-from airflow.providers.amazon.aws.hooks.base_aws import AwsBaseHook
-from airflow.providers.amazon.aws.operators.sagemaker_base import SageMakerBaseOperator
+"""This module is deprecated. Please use :mod:`airflow.providers.amazon.aws.operators.sagemaker`."""
 
+import warnings
 
-class SageMakerTransformOperator(SageMakerBaseOperator):
-    """
-    Initiate a SageMaker transform job.
+from airflow.providers.amazon.aws.operators.sagemaker import SageMakerTransformOperator  # noqa
 
-    This operator returns The ARN of the model created in Amazon SageMaker.
-
-    :param config: The configuration necessary to start a transform job (templated).
-
-        If you need to create a SageMaker transform job based on an existed SageMaker model::
-
-            config = transform_config
-
-        If you need to create both SageMaker model and SageMaker Transform job::
-
-            config = {
-                'Model': model_config,
-                'Transform': transform_config
-            }
-
-        For details of the configuration parameter of transform_config see
-        :py:meth:`SageMaker.Client.create_transform_job`
-
-        For details of the configuration parameter of model_config, See:
-        :py:meth:`SageMaker.Client.create_model`
-
-    :type config: dict
-    :param aws_conn_id: The AWS connection ID to use.
-    :type aws_conn_id: str
-    :param wait_for_completion: Set to True to wait until the transform job finishes.
-    :type wait_for_completion: bool
-    :param check_interval: If wait is set to True, the time interval, in seconds,
-        that this operation waits to check the status of the transform job.
-    :type check_interval: int
-    :param max_ingestion_time: If wait is set to True, the operation fails
-        if the transform job doesn't finish within max_ingestion_time seconds. If you
-        set this parameter to None, the operation does not timeout.
-    :type max_ingestion_time: int
-    """
-
-    def __init__(
-        self,
-        *,
-        config: dict,
-        wait_for_completion: bool = True,
-        check_interval: int = 30,
-        max_ingestion_time: Optional[int] = None,
-        **kwargs,
-    ):
-        super().__init__(config=config, **kwargs)
-        self.config = config
-        self.wait_for_completion = wait_for_completion
-        self.check_interval = check_interval
-        self.max_ingestion_time = max_ingestion_time
-        self.create_integer_fields()
-
-    def create_integer_fields(self) -> None:
-        """Set fields which should be casted to integers."""
-        self.integer_fields: List[List[str]] = [
-            ['Transform', 'TransformResources', 'InstanceCount'],
-            ['Transform', 'MaxConcurrentTransforms'],
-            ['Transform', 'MaxPayloadInMB'],
-        ]
-        if 'Transform' not in self.config:
-            for field in self.integer_fields:
-                field.pop(0)
-
-    def expand_role(self) -> None:
-        if 'Model' not in self.config:
-            return
-        config = self.config['Model']
-        if 'ExecutionRoleArn' in config:
-            hook = AwsBaseHook(self.aws_conn_id, client_type='iam')
-            config['ExecutionRoleArn'] = hook.expand_role(config['ExecutionRoleArn'])
-
-    def execute(self, context) -> dict:
-        self.preprocess_config()
-
-        model_config = self.config.get('Model')
-        transform_config = self.config.get('Transform', self.config)
-
-        if model_config:
-            self.log.info('Creating SageMaker Model %s for transform job', model_config['ModelName'])
-            self.hook.create_model(model_config)
-
-        self.log.info('Creating SageMaker transform Job %s.', transform_config['TransformJobName'])
-        response = self.hook.create_transform_job(
-            transform_config,
-            wait_for_completion=self.wait_for_completion,
-            check_interval=self.check_interval,
-            max_ingestion_time=self.max_ingestion_time,
-        )
-        if response['ResponseMetadata']['HTTPStatusCode'] != 200:
-            raise AirflowException(f'Sagemaker transform Job creation failed: {response}')
-        else:
-            return {
-                'Model': self.hook.describe_model(transform_config['ModelName']),
-                'Transform': self.hook.describe_transform_job(transform_config['TransformJobName']),
-            }
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.operators.sagemaker`.",
+    DeprecationWarning,
+    stacklevel=2,
+)
