+++ airflow/providers/amazon/aws/sensors/sagemaker_tuning.py
@@ -16,39 +16,14 @@
 # specific language governing permissions and limitations
 # under the License.
 
-from airflow.providers.amazon.aws.hooks.sagemaker import SageMakerHook
-from airflow.providers.amazon.aws.sensors.sagemaker_base import SageMakerBaseSensor
+"""This module is deprecated. Please use :mod:`airflow.providers.amazon.aws.sensors.sagemaker`."""
 
+import warnings
 
-class SageMakerTuningSensor(SageMakerBaseSensor):
-    """
-    Asks for the state of the tuning state until it reaches a terminal state.
-    The sensor will error if the job errors, throwing a AirflowException
-    containing the failure reason.
+from airflow.providers.amazon.aws.sensors.sagemaker import SageMakerTuningSensor  # noqa
 
-    :param job_name: job_name of the tuning instance to check the state of
-    :type job_name: str
-    """
-
-    template_fields = ['job_name']
-    template_ext = ()
-
-    def __init__(self, *, job_name: str, **kwargs):
-        super().__init__(**kwargs)
-        self.job_name = job_name
-
-    def non_terminal_states(self):
-        return SageMakerHook.non_terminal_states
-
-    def failed_states(self):
-        return SageMakerHook.failed_states
-
-    def get_sagemaker_response(self):
-        self.log.info('Poking Sagemaker Tuning Job %s', self.job_name)
-        return self.get_hook().describe_tuning_job(self.job_name)
-
-    def get_failed_reason_from_response(self, response):
-        return response['FailureReason']
-
-    def state_from_response(self, response):
-        return response['HyperParameterTuningJobStatus']
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.sagemaker`.",
+    DeprecationWarning,
+    stacklevel=2,
+)
